Convert forceApexTriggerCreate
