+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/forceApexTriggerCreate.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { expect } from 'chai';
-import { ForceApexTriggerCreateExecutor } from '../../../src/commands/forceApexTriggerCreate';
+import { ForceApexTriggerCreateExecutor } from '../../../src/commands/templates/forceApexTriggerCreate';
 import { nls } from '../../../src/messages';
 
 // tslint:disable:no-unused-expression
